RF02-TC-046: reorganize structure
