--- src/main/java/org/umssdiplo/automationv01/core/managepage/Position/EditPosition.java
@@ -1,4 +1,0 @@
-package org.umssdiplo.automationv01.core.managepage.Position;
-
-public class EditPosition {
-}
