--- src/main/java/org/umssdiplo/automationv01/core/managepage/Position/ListPosition.java
@@ -1,16 +1,0 @@
-package org.umssdiplo.automationv01.core.managepage.Position;
-
-import cucumber.api.java.After;
-import cucumber.api.java.Before;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-import org.umssdiplo.automationv01.core.customwebdriver.ManageDriver;
-import org.umssdiplo.automationv01.core.utils.CommonEvents;
-
-import java.util.List;
-
-public class ListPosition {
-
-
-
-}
