+++ src/main/java/org/umssdiplo/automationv01/core/utils/LoadPage.java
@@ -2,9 +2,7 @@ package org.umssdiplo.automationv01.core.utils;
 
 import org.umssdiplo.automationv01.core.customwebdriver.ManageDriver;
 import org.umssdiplo.automationv01.core.managepage.Login.Login;
-import org.umssdiplo.automationv01.core.managepage.Position.CreatePosition;
-import org.umssdiplo.automationv01.core.managepage.Position.EditPosition;
-import org.umssdiplo.automationv01.core.managepage.Position.ListPosition;
+import org.umssdiplo.automationv01.core.managepage.Position.CreateEditPosition;
 import org.umssdiplo.automationv01.core.managepage.Position.Position;
 import org.umssdiplo.automationv01.core.managepage.Util;
 
@@ -23,15 +21,7 @@ public final class LoadPage {
         return new Position();
     }
 
-    public static ListPosition listPositions() {
-        return new ListPosition();
-    }
-
-    public static CreatePosition createPositionPage() {
-        return new CreatePosition();
-    }
-
-    public static EditPosition editPositionPage() {
-        return new EditPosition();
+    public static CreateEditPosition createPositionPage() {
+        return new CreateEditPosition();
     }
 }
